fix/feat: Corrigir Facade
